requirements.txt
Faker>=37.1.0
fastapi>=0.100.0
httpx>=0.27.0
uvicorn[standard]>=0.20.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
unstructured[local-inference]>=0.12.0
python-dotenv>=1.0.0
pytest>=8.3.5
pytest-mock>=3.14.0
timescale_vector>=0.1.0
psycopg2-binary
ollama>=0.1.0
pandas>=1.5.0
